Fix where and how pane appears
The pane always appeared on the main space,
now it appears on the same space as Alfred.

The pane appeared on top of other windows,
but not on top of `iTerm2`'s visor window.
That's because `iTerm2` uses the `screenSaver`
window level. Now we start using it too. Thus
making the pane no longer appear behind the
iterm window.