Fix bug preventing usage of query params in file URLs
We can create HTML files whose content depends on
query params. Then we can set quicklookurl to
`file:///path/to/file.html?param=value`. This didn't
work so far because the URL string didn't end in `.html`
and thus, the extra pane decided to not render it at all.
We fix this bug by considering only the "path" component
of the URL to check whether it is an HTML file.