Hide all panes on selection change before showing relevant panes
Consider two workflows, one provides `def` keyword for
dictionary lookup and another that provides `gdef` keyword
for searching Google for definition.

Consider the sequence of user actions and pane behavior:
1. User types "def blah"
2. We show the dictionary pane
3. User changes the typed query to "gdef blah"
4. We show the google search pane

Note that at 4, we now have two panes visible as pane from
2 was never hidden.

With this change, we add another step between 2 and 3 for
hiding the dictionary pane.